Moved platform/input/element-input to /framework/input
