+++ src/framework/app-base.js
@@ -70,7 +70,7 @@ import {
 
 /** @typedef {import('../platform/graphics/graphics-device.js').GraphicsDevice} GraphicsDevice */
 /** @typedef {import('../platform/graphics/texture.js').Texture} Texture */
-/** @typedef {import('../platform/input/element-input.js').ElementInput} ElementInput */
+/** @typedef {import('./input/element-input.js').ElementInput} ElementInput */
 /** @typedef {import('../platform/input/game-pads.js').GamePads} GamePads */
 /** @typedef {import('../platform/input/keyboard.js').Keyboard} Keyboard */
 /** @typedef {import('../platform/input/mouse.js').Mouse} Mouse */
