+++ src/framework/app-options.js
@@ -1,6 +1,6 @@
 /** @typedef {import('../framework/handlers/handler.js').ResourceHandler} ResourceHandler */
 /** @typedef {import('../platform/graphics/graphics-device.js').GraphicsDevice} GraphicsDevice */
-/** @typedef {import('../platform/input/element-input.js').ElementInput} ElementInput */
+/** @typedef {import('./input/element-input.js').ElementInput} ElementInput */
 /** @typedef {import('../platform/input/game-pads.js').GamePads} GamePads */
 /** @typedef {import('../platform/input/keyboard.js').Keyboard} Keyboard */
 /** @typedef {import('../platform/input/mouse.js').Mouse} Mouse */
